BIM: move more functions to their own modules, standardize logging
